Refactor generate tx
