+++ src/commands/publish/handler.ts
@@ -1,19 +1,10 @@
 import Listr from "listr";
-import {
-  defaultComposeFileName,
-  defaultDir,
-  YargsError
-} from "../../params.js";
-import {
-  ListrContextBuildAndPublish,
-  ReleaseType,
-  releaseTypes
-} from "../../types.js";
+import { defaultComposeFileName, defaultDir } from "../../params.js";
+import { ListrContextBuildAndPublish } from "../../types.js";
 import { VerbosityOptions } from "../build/types.js";
 import { PublishCommandOptions } from "./types.js";
 import { publish } from "../../tasks/publish/index.js";
-
-const typesList = releaseTypes.join(" | ");
+import { parseReleaseType } from "./parseReleaseType.js";
 
 /**
  * Common handler for CLI and programatic usage
@@ -80,46 +71,4 @@ export async function publishHandler({
   );
 
   return await publishTasks.run();
-}
-
-function parseReleaseType({ type }: { type?: string }): ReleaseType {
-  const tag = process.env.TRAVIS_TAG || process.env.GITHUB_REF;
-  const typeFromEnv = process.env.RELEASE_TYPE;
-
-  /**
-   * Custom options to pass the type argument
-   */
-  if (!type) {
-    if (typeFromEnv) type = typeFromEnv;
-
-    if (tag?.includes("release")) type = parseReleaseTypeFromTag(tag);
-  }
-
-  if (!type && typeFromEnv) {
-    type = typeFromEnv as ReleaseType;
-  }
-  if (!type && tag?.includes("release")) {
-    type = parseReleaseTypeFromTag(tag);
-  }
-
-  validateReleaseType(type);
-
-  return type as ReleaseType;
-}
-
-function parseReleaseTypeFromTag(tag: string): ReleaseType {
-  return (tag.split("release/")[1] || "patch") as ReleaseType;
-}
-
-/**
- * Make sure the release type exists and is correct
- */
-function validateReleaseType(type?: string) {
-  if (!type)
-    throw new YargsError(`Missing required argument [type]: ${typesList}`);
-
-  if (!releaseTypes.includes(type as ReleaseType))
-    throw new YargsError(
-      `Invalid release type "${type}", must be: ${typesList}`
-    );
 }
