+++ src/tasks/generatePublishTx/types.ts
@@ -1,0 +1,3 @@
+import { ApmRepository } from "@dappnode/toolkit";
+
+export type Repo = Awaited<ReturnType<ApmRepository["getRepoContract"]>>;
